% Generated by roxygen2 (4.0.1): do not edit by hand
\name{liveWatApp}
\alias{liveWatApp}
\title{Create a new App for live assisted teaching}
\usage{
liveWatApp(wat, refresh.time = 2000L)
}
\arguments{
\item{wat}{a wat object}

\item{refresh.time}{time interval in which clients will be refreshed in milli-seconds}
}
\description{
Create a new App for live assisted teaching
}

