% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simpleTeachingApp}
\alias{simpleTeachingApp}
\title{An example of a simple teaching app with some given default questions in German}
\usage{
simpleTeachingApp(run.app = FALSE)
}
\description{
An example of a simple teaching app with some given default questions in German
}

